removed plus part
